+++ CHANGELOG.md
@@ -30,3 +30,4 @@
 * Add `Image::push`
 * Add `Images::prune`
 * Add `Images::search`
+* Add `Images::export`